tests: refactor cli tests
