+++ tests/functional/cli/storage/test_container.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2016 OpenIO SAS
+# Copyright (C) 2016-2017 OpenIO SAS
 
 # This library is free software; you can redistribute it and/or
 # modify it under the terms of the GNU Lesser General Public
@@ -12,10 +12,10 @@
 # License along with this library.
 
 import uuid
-from tests.functional import TestCase
+from tests.functional.cli import CliTestCase
 
 
-class ContainerTest(TestCase):
+class ContainerTest(CliTestCase):
     """Functional tests for containers."""
     NAME = uuid.uuid4().hex
